backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-openai
python-dotenv
ddgs
tavily-python
